Gendollardollar/AiTradeSignal/GenX/python/requirements.txt

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
schedule>=1.2.0
joblib>=1.3.0

# Trading and market data
python-telegram-bot>=20.0
ccxt>=4.0.0
ta-lib>=0.4.28
yfinance>=0.2.18

# AI and machine learning
openai>=1.0.0
google-generativeai>=0.3.0
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.30.0

# Discord integration
discord.py>=2.3.0
aiohttp>=3.8.0

# Database and caching
psycopg2-binary>=2.9.0
redis>=4.5.0
sqlalchemy>=2.0.0

# Web framework for API endpoints
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0

# Data processing and analysis
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
websocket-client>=1.6.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
